incomplete code

+++ scripts/script1.js
@@ -12,14 +12,14 @@ class AddressBook {
 
     printAll(info) {
         let contactsLen = this.contacts.length;
-        for (let i = 0; i < contactsLen; i++) {            
+        for (let i = 0; i < contactsLen; i++) {
             console.log(this.contacts[i]);
         }
     }
 
     deleteAt(index) {
         index = prompt("Enter Index Number To Delete ")
-        this.contacts.splice(index,1);
+        this.contacts.splice(index, 1);
     }
 
     // upDate();
@@ -32,7 +32,7 @@ class Contact {
         this.email = email;
         this.relation = relation;
     }
-  
+
 }
 
 const book = new AddressBook();
@@ -40,22 +40,54 @@ book.add(new Contact("Jonathon", "3137892345", "jonB@yahoo.com", "Myself"));
 book.add(new Contact("Josh", "3135670098", "joshC@aol.com", "friend"))
 console.log(book);
 
-do{
-    const select = prompt("Would You Like To (A)dd, (D)elete, (P)rint Or (Q)uit? (A/D/P/Q)");
-    switch (select.toLocaleLowerCase()){
-        case "a": this.name = newName;
-                  this.phone = newPhone;
-                  this.email = newEmail;
-                  this.relation = newRelation;
-                  newName = prompt("Enter New Contact (Name).");
-                  newPhone = prompt("Enter New Contact (Phone).");
-                  newEmail = prompt("Enter New Contact (Email).");
-                  newRelation = prompt("Enter New Contact (Relation).");
-                  add();
+ do{
+   let select = prompt("Would You Like To (A)dd, (D)elete, (P)rint Or (Q)uit? (A/D/P/Q)");
 
-    }
+   if(select.toLocaleLowerCase() === "a"){
+    this.name = prompt("Enter New Contact (Name).");
+    this.phone = prompt("Enter New Contact (Phone).");
+    this.email = prompt("Enter New Contact (Email).");
+    this.relation = prompt("Enter New Contact (Relation).");
+    book.add();
+   }
+   if(select.toLocaleLowerCase() === "d"){
+    this.index = prompt("Enter Index Of Contact To Delete.");
+    book.deleteAt();
+   }
+   
+   if(select.toLocaleLowerCase() === "p"){
+    book.printAll();
+   }
+   if(select.toLocaleLowerCase() === "q"){
+       break;
+   }
+   else if(select.toLocaleLowerCase() === " "){
+        console.log("Not A Valid Entry.");
+   }
 }
-while(select !== "Q")
-book.printAll();
-book.deleteAt();
-book.printAll();
+while (select !== "q");
+//     switch (select.toLocaleLowerCase()) {
+//         case "a":
+//             //   this.name = newName;
+//             //   this.phone = newPhone;
+//             //   this.email = newEmail;
+//             //   this.relation = newRelation;
+//             this.name = prompt("Enter New Contact (Name).");
+//             this.phone = prompt("Enter New Contact (Phone).");
+//             this.email = prompt("Enter New Contact (Email).");
+//             this.relation = prompt("Enter New Contact (Relation).");
+//             book.add();
+//         //    let select = prompt("Would You Like To (A)dd, (D)elete, (P)rint Or (Q)uit? (A/D/P/Q)");
+//         case "d":
+//             this.index = prompt("Enter Index Of Contact To Delete.");
+//             book.deleteAt();
+
+//         case "q":
+//             break;
+//     }
+// }
+
+
+// book.printAll();
+// book.deleteAt();
+// book.printAll();
